sr(api): move trainings controller in devcomp
