+++ api/lib/routes.js
@@ -45,7 +45,6 @@ import * as sessions from './application/sessions/index.js';
 import * as tags from './application/tags/index.js';
 import * as targetProfiles from './application/target-profiles/index.js';
 import * as targetProfilesManagement from './application/target-profiles-management/index.js';
-import * as trainings from './application/trainings/index.js';
 import * as missions from './application/missions/index.js';
 import * as frameworks from './application/frameworks/index.js';
 import * as tutorialEvaluations from './application/tutorial-evaluations/index.js';
@@ -102,7 +101,6 @@ const routes = [
   targetProfiles,
   missions,
   targetProfilesManagement,
-  trainings,
   frameworks,
   tutorialEvaluations,
   userOrgaSettings,
