+++ api/src/devcomp/routes.js
@@ -1,5 +1,6 @@
 import * as modulesRoutes from './application/modules/index.js';
+import * as trainings from './application/trainings/index.js';
 
-const devcompRoutes = [modulesRoutes];
+const devcompRoutes = [modulesRoutes, trainings];
 
 export { devcompRoutes };
