+++ api/tests/devcomp/unit/application/trainings/training-controller_test.js
@@ -1,8 +1,9 @@
-import { sinon, expect, hFake } from '../../../test-helper.js';
-import { trainingController } from '../../../../lib/application/trainings/training-controller.js';
-import { usecases } from '../../../../lib/domain/usecases/index.js';
-import { TrainingTrigger } from '../../../../src/devcomp/domain/models/TrainingTrigger.js';
-import { DomainTransaction } from '../../../../lib/infrastructure/DomainTransaction.js';
+import { sinon, expect, hFake } from '../../../../test-helper.js';
+import { trainingController } from '../../../../../src/devcomp/application/trainings/training-controller.js';
+import { usecases } from '../../../../../src/devcomp/domain/usecases/index.js';
+import { usecases as libUsecases } from '../../../../../lib/domain/usecases/index.js';
+import { TrainingTrigger } from '../../../../../src/devcomp/domain/models/TrainingTrigger.js';
+import { DomainTransaction } from '../../../../../lib/infrastructure/DomainTransaction.js';
 
 describe('Unit | Controller | training-controller', function () {
   describe('#findPaginatedTrainingSummaries', function () {
@@ -314,7 +315,7 @@ describe('Unit | Controller | training-controller', function () {
       const trainingId = 145;
       const targetProfileSummaries = Symbol('targetProfileSummaries');
       const serializedTargetProfileSummaries = Symbol('serializedTargetProfileSummaries');
-      sinon.stub(usecases, 'findTargetProfileSummariesForTraining').resolves(targetProfileSummaries);
+      sinon.stub(libUsecases, 'findTargetProfileSummariesForTraining').resolves(targetProfileSummaries);
 
       const targetProfileSummaryForAdminSerializer = {
         serialize: sinon.stub(),
@@ -329,7 +330,7 @@ describe('Unit | Controller | training-controller', function () {
       });
 
       // then
-      expect(usecases.findTargetProfileSummariesForTraining).to.have.been.calledWithExactly({ trainingId });
+      expect(libUsecases.findTargetProfileSummariesForTraining).to.have.been.calledWithExactly({ trainingId });
       expect(result).to.be.equal(serializedTargetProfileSummaries);
     });
   });
